+++ lambda/test/package.json
@@ -12,12 +12,11 @@
     "test": "sls invoke --function chromium-test --log"
   },
   "dependencies": {
-    "chalk": "^5.3.0",
     "puppeteer-core": "^19.6.3",
     "winston": "^3.14.2"
   },
   "devDependencies": {
-    "@sparticuz/chromium": "110.0.0",
+    "@sparticuz/chromium": "^123.0.0",
     "serverless": "^3.27.0"
   }
 }
